add Courier class and its support, add new package types

+++ Human.cpp
@@ -18,7 +18,7 @@ int Human::getBla()
     return bla;
 }
 
-void Human::recharge(int ch)
+void Human::recharge(double ch)
 {
     bla += ch;
 }
